requirements.txt
playwright==1.49.1
pytest==8.3.4
pytest-playwright==0.6.2
allure-pytest==2.15.3
allure-python-commons==2.15.3
